Fix: Posts published while feediverse is running are not tooted.
Fix race-condition: If a post was published within the short period of
time between fetching the RSS feed and saving the config-file, this
post was not tooted. This was caused by the timestamp in the
config-file having been the time when the file was written, not when
the feed was fetched.

I (hopefully) fixed this by storing the latest post's timestamp in the
config file. This still might cause the same issue if several feeds
are checked using the same config file.